Add detect() routine
The timing for writing to the slave unit seemed to be off (writing
seemed to have no effect, so we figured that it could be caused by us
having assumed the wrong divider for the slave unit), so we implemented
the detect() routine used by NESizer2. This showed, much to our
surprise, that the slave unit seems to be running on a divider of 15,
and with this the timing seems to work.